LPD-30649 Extract RevisionHistory to its own component and convert it to Typescript
